led显示
工程中led显示功能，led数据通过485通信后，数据同步更新。

+++ AF1_lib/HARDWARE/RS485/rs485.h
@@ -14,7 +14,7 @@
 //All rights reserved									  
 //////////////////////////////////////////////////////////////////////////////////
 	  		  	
-extern u8 RS485_RX_BUF[64]; 		//接收缓冲,最大64个字节
+extern u16 RS485_RX_BUF[64]; 		//接收缓冲,最大64个字节
 extern u8 RS485_RX_CNT;   			//接收到的数据长度
 
 //模式控制
@@ -27,7 +27,7 @@ extern u8 RS485_RX_CNT;   			//接收到的数据长度
 
 void RS485_Init(u32 bound);
 void RS485_Send_Data(u8 *buf,u8 len);
-void RS485_Receive_Data(u8 *buf,u8 *len);
+void RS485_Receive_Data(u32 *buf,u8 *len);
 
 
 #endif
